Upgraded Cmake files to handle tests.
To add a test, create a new file under tests/,
add an add_subdirectory line to tests/CMakeLists.txt,
Create a CMakeLists.txt in the directory for the particular test.
TODO: create a code to run all the tests at once and provide a method
to determine success or failure.