Updated tests. Creating code for feedforward networks.
Realized I needed 3D tensors to handle dimensions, batch and sequence length.
Thus, we will move to armadillo (cannot handle CUDA thought). So we stick to CPU
implemenation for now. later, we can switch back to eigen that has cuda support (although
this involves individual matrix multiplies on CUDA, which will only be efficient if it
possible to leave the results of calculations on the GPU.