Backprop (weights) works with softmax if network has no recursion.
Bias do not work yet. Difficulty is Jacobian, which is computed on the fly.
The Jacobian is computed on the fly. If seq_len = 16 and batch_size=32 and
layer_size=128, the Jacobian would have size 16*32*128 = 65,000 (I guess I could
store it). It would be cheaper since I need to call it twice during backprop.