Fixed error in input variables to network when seq_len > 1.
Also fixed some problems with temporal connections in print_utils.
alpha now converges to a fixed value for larger sequences. But if I
increase seq_len by 1, the converged alpha also increases by 1. Why woudl
that be? What I do not do is take the average of the loss function with respect
to the sequence length. Could this be it? I doubt it as it would simply be
equivalent to changing the learning rate (which after checking, does not
influence the converged alpha).